test: share one configurable notify fake
Both files stood in for createNotification() with a hoisted daemon gate
and the same explanatory comment about the D-Bus probe, then diverged:
one built a real electron Notification, the other recorded a bespoke
object. The shared fake offers both modes and lets a test observe the
force argument the Last.fm connect-failure path passes.
